+++ server/routes/mongoRoutes.js
@@ -1,5 +1,5 @@
 const express = require('express');
-const db = require('./database/index');
+const db = require('../database/mongoDB');
 
 const router = express.Router();
 
@@ -9,25 +9,25 @@ router.use(express.urlencoded());
 router.post('/', (req, res) => {
   db.createItem(req.body.images)
     .then((rows) => res.status(200).json(rows))
-    .catch((err) => res.status(400).json({ message: `Error: ${err}` }));
+    .catch((err) => res.status(500).json(err));
 });
 
 router.get('/:id', (req, res) => {
   db.getItem(req.params.id)
     .then((rows) => res.status(200).json(rows))
-    .catch((err) => res.status(400).json({ message: `Error: ${err}` }));
+    .catch((err) => res.status(500).json(err));
 });
 
 router.put('/:id', (req, res) => {
   db.updateItem(req.params.id, req.params.imageIndex, req.params.image)
     .then((rows) => res.status(200).json(rows))
-    .catch((err) => res.status(400).json({ message: `Error: ${err}` }));
+    .catch((err) => res.status(500).json(err));
 });
 
 router.delete('/:id', (req, res) => {
   db.deleteItem(req.params.id)
     .then((rows) => res.status(200).json(rows))
-    .catch((err) => res.status(400).json({ message: `Error: ${err}` }));
+    .catch((err) => res.status(500).json(err));
 });
 
 module.exports = router;
