--- server/test.yml
@@ -1,32 +1,0 @@
-config:
-  target: 'http://localhost:3001'
-  processor: "./stressTest.js"
-  phases:
-    - duration: 180
-      arrivalRate: 1
-      rampTo: 1000
-      maxVusers: 1000
-      name: "Ramp up from 1 to 1k VUs per second"
-    - pause: 60
-    - duration: 360
-      arrivalRate: 1000
-      name: "Sustained max load"
-
-scenarios:
-  - name: "GET product images"
-    weight: 9
-    flow:
-      - function: "getRandom"
-      - get:
-          url: "/api/items/{{itemId}}"
-      - log: "Successful GET request to id: {{itemId}}"
-  - name: "POST product images"
-    weight: 1
-    flow:
-      - function: "postRandom"
-      - post:
-          url: "/api/items"
-          json:
-            itemId: "{{itemId}}"
-            images: "{{images}}"
-      - log: "Successful POST request to itemId: {{itemId}}"
